refactor/#473: Review페이지 경로 수정
